.\cordova does not work, needs to be %~dp0\cordova
The ".\cordova" construction only works if you are in the .cvm directory when you run this script. The problem is that "." is interpreted as the current directory from where you issued the command, it does not represent where the script file was found. Assuming "cordova.cmd" is located in %HOMEPATH%\.cvm (and assuming your PATH is setup correctly), if you issue the "cordova" command from anywhere but the %HOMEPATH%\.cvm directory, the ".\cordova" construct will fail.